backend/data/src/main/kotlin/io/tolgee/formats/json/out/JsonFileExporter.kt
package io.tolgee.formats.json.out

import com.fasterxml.jackson.databind.ObjectMapper
import io.tolgee.dtos.IExportParams
import io.tolgee.formats.ExportFormat
import io.tolgee.formats.genericStructuredFile.out.GenericStructuredFileExporter
import io.tolgee.formats.nestedStructureModel.StructureModelBuilder
import io.tolgee.service.export.dataProvider.ExportTranslationView
import io.tolgee.service.export.exporters.FileExporter
import java.io.InputStream

class JsonFileExporter(
  override val translations: List<ExportTranslationView>,
  override val exportParams: IExportParams,
  private val projectIcuPlaceholdersSupport: Boolean,
  val objectMapper: ObjectMapper,
) : FileExporter {
  override val fileExtension: String = ExportFormat.JSON.extension

  private val genericExporter =
    GenericStructuredFileExporter(
      translations = translations,
      exportParams = exportParams,
      fileExtension = fileExtension,
      projectIcuPlaceholdersSupport = projectIcuPlaceholdersSupport,
      objectMapper = objectMapper,
      rootKeyIsLanguageTag = false,
      pluralsViaNesting = false,
      placeholderConvertorFactory = null,
    )

  val result: LinkedHashMap<String, StructureModelBuilder> = LinkedHashMap()

  override fun produceFiles(): Map<String, InputStream> {
    return genericExporter.produceFiles()
  }
}
